helper_functions.py - renamed to missing_file_checker_functions.py and consequential changes in other files relying on it
